Initial implementation of events triggered by turn start and turn end
