+++ .eslintrc.json
@@ -74,7 +74,6 @@
         "nonwords": false
       }
     ],
-    "no-prototype-builtins": "off",
-    "no-return-await": "error"
+    "no-prototype-builtins": "off"
   }
 }
