+++ src/lang/en.json
@@ -1523,6 +1523,7 @@
                 "DamageNotes": "Damage Notes",
                 "DamageNotesTooltip": "The damage notes are displayed on the damage roll chat card. This is a great way to put helpful reminders, especially for GMs with complicated NPCs.",
                 "DamageSection": "Damage Section {section}",
+                "DeleteDamageSection": "Delete damage sSection",
                 "Details": "Action Details",
                 "HealingFormula": "Healing Formula",
                 "OtherFormula": "Other Formula",
@@ -2935,6 +2936,24 @@
         "Touch": "Touch",
         "TraitArmorProf": "Armor Proficiencies",
         "TraitWeaponProf": "Weapon Proficiencies",
+        "TurnEvent": {
+          "TurnEventI": "Turn Event {number}",
+          "AddTurnEvent": "Add Turn Event",
+          "DeleteTurnEvent": "Delete Turn Event",
+          "Name": "Name",
+          "Type": "Type",
+          "TypeTooltip": "The type of behaviour to trigger for this turn event.",
+          "Trigger": "Trigger",
+          "TriggerTooltip": "When during a combatant's turn to trigger the event.",
+          "Formula": "Formula",
+          "FormulaTooltip": "A formula which will be rolled when the trigger is met.",
+          "NoteContent": "Note Content",
+          "NoteTooltip": "Text which will be sent to chat as a chat message when the trigger is met.",
+          "Types": {
+            "Roll": "Roll",
+            "Note": "Note"
+          }
+        },
         "UPBs": "UPB's",
         "Units": {
             "Speed": "ft."
