+++ src/less/items.less
@@ -607,6 +607,7 @@
     .item-duration {
         .origin {
             float: left;
+            position: static;
 
             &.item-properties {
                 img {
