+++ src/less/variables.less
@@ -194,7 +194,7 @@
             }
 
             // 2nd nested bubble
-            .bubble .bubble-info {
+            .bubble {
                 border: 2px solid #7a1515;
 
                 .bubble-header {
