+++ src/template.json
@@ -2626,7 +2626,8 @@
       "type": "effect",
       "enabled": true,
       "context": {},
-      "showOnToken": true
+      "showOnToken": true,
+      "turnEvents": []
     }
   }
 }
