+++ src/templates/items/parts/starship-subactions.hbs
@@ -43,7 +43,7 @@
                 </div>
 
                 {{!-- Subaction Effects --}}
-                {{#with (lookup ../editorInfo id) as |editorInfo|}}
+                {{#with (lookup ../subactionEditorInfo id) as |editorInfo|}}
                     <div class="bubble">
                         <h3 class="bubble-header">{{ localize "SFRPG.ItemSheet.StarshipAction.NormalEffect" }}</h3>
